+++ core/parser/src/main/java/com/blazebit/persistence/impl/expression/NumericLiteral.java
@@ -74,5 +74,9 @@ public class NumericLiteral extends AbstractNumericExpression {
         return result;
     }
 
+    @Override
+    public String toString() {
+        return value;
+    }
 }